requirements.txt
Django==4.2.11
djangorestframework==3.15.1
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.9  # For PostgreSQL, optional if using SQLite